rename slice
